requirements.txt
alembic==1.8.0
anyio==4.9.0
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
blinker==1.9.0
Brotli==1.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
ConfigArgParse==1.7
coverage==7.8.0
fastapi==0.115.12
fastapi-utils==0.8.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
h11==0.13.0
httpcore==1.0.7
httpx==0.28.1
idna==3.3
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.33.2
lxml==4.9.1
Mako==1.2.1
MarkupSafe==2.1.1
msgpack==1.1.0
outcome==1.2.0
packaging==24.2
pluggy==1.5.0
psutil==5.9.8
psycopg2-binary==2.9.3
pydantic==1.10.0
PyJWT==2.10.1
PySocks==1.7.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.20
pyzmq==26.3.0
requests==2.28.1
selenium==4.4.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.29
starlette==0.46.1
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.13.0
urllib3==1.26.12
uvicorn==0.18.2
webdriver-manager==3.8.3
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
